requirements.txt
alabaster==0.7.7
amqp==2.2.2
apturl==0.5.2
audioread==2.1.5
Babel==1.3
beautifulsoup4==4.4.1
billiard==3.5.0.3
bleach==2.1.2
boto==2.48.0
boto3==1.5.8
botocore==1.8.22
Brlapi==0.6.4
bz2file==0.98
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
chrome-gnome-shell==0.0.0
click==6.7
command-not-found==0.3
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
Django==1.11.6
django-extensions==1.9.6
django-rest-framework==0.1.0
djangorestframework==3.7.1
docutils==0.14
entrypoints==0.2.3
filelock==2.0.13
gensim==3.2.0
html5lib==1.0.1
idna==2.6
ipykernel==4.7.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.0
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
joblib==0.11
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.1
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==4.1.0
language-selector==0.1
librosa==0.5.1
llvmlite==0.20.0
louis==2.6.4
lxml==3.5.0
Mako==1.0.3
Markdown==2.6.6
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.8.3
natsort==5.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
notebook==5.2.2
numba==0.35.0
numpy==1.13.3
oauthlib==2.0.6
pandas==0.20.3
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.3.1
pickleshare==0.7.4
Pillow==3.1.2
prompt-toolkit==1.0.15
psycopg2==2.7.3.1
ptyprocess==0.5.2
py==1.4.34
pycups==1.9.73
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pymongo==3.5.1
pyparsing==2.2.0
pyspotlight==0.7.1
pytest==3.2.3
pytest-django==3.1.2
python-apt==1.1.0b1
python-dateutil==2.6.1
python-debian==0.1.27
python-systemd==231
pytz==2017.2
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.3
qtconsole==4.3.1
raven==6.2.1
remarkable==1.87
reportlab==3.3.0
request==0.0.22
requests==2.18.4
requests-oauthlib==0.8.0
resampy==0.2.0
roman==2.0.0
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.6
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
system-service==0.3
terminado==0.8.1
testpath==0.3.1
torch==0.2.0.post3
tornado==4.5.3
traitlets==4.3.2
tweepy==3.5.0
ubuntu-drivers-common==0.0.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.0
vine==1.1.4
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.1.0
xdiagnose==3.8.4.1
xkit==0.0.0
